requirements.txt
streamlit
pandas
pymongo
pymongo[srv]
boto
python-decouple
plotly_express
datetime

